requirements.txt
uliweb-alembic
fastapi
strawberry-graphql[fastapi]>=0.205.0
sqlmodel
sqlalchemy
alembic
psycopg2-binary  # Postgres driver
uvicorn[standard]
python-dotenv
pytest
httpx>=0.28.0
pytest-asyncio
typer
pydantic-settings
asyncpg
pydantic